fix: unpdated code using unsafe library
The usage of sha1 was removed and updated to sha256. The use of `sha*`
is minor compared to the changes which fix the functionality of creating
unique k8s object names based on the LeaseID.

Added test to assert that the name is properly passed into the k8s
type.

fixes #550

Signed-off-by: Josh Roppo <josh@akash.network>