refactor: provider test closures
golangci lint'er does not appreciate the nested anonymous function calls
using an anonymous `test` struct. This refactor creates functions to
produce the func(t *testing.T) interface so that functionality of the
tests can still be nested for clean output, but results in more complex
code. However I do appreciate the encapsulation of the test
functionality with the types.

Signed-off-by: Joshua Roppo <josh@akash.network>