feature: Update 'kube' example for akash network (#612)
* feature: Update 'kube' example for akash network

* Makefile to execute akash* built binaries with consistency.
`akashctl provider` command re-enabled.
* Debugging and testing code around read provider config file.
* Functional `...tx send...<from> <to> <value>` command added as
`send-to-main`. Additional `query-txn`
* README added. Functional querying of market information.
* Makefile to execute akash* built binaries with consistency.
`akashctl provider` command re-enabled.
* Removing more differences betweene lite and kube _run scripts
* Identified Provider creation missing validation
* Fixed: Empty kv pairs could be applied to Provider Create and Update Messages.
This results in a Tendermint error, which is a big red herring to the
actual problem. Discovered via YAML configuration being parsed as empty.
* Unit tests added to verify the `provider` module's Msg validation
code.
* `make run-provider` command added.
* Documenting running Provider and Bidding
* Allow Provider spec w/o attributes
  * @boz requested this change because requiring Providers to have
attributes is not part of the spec. It will enable errors to be returned
from the Tendermint layer because that does require attributes for some
reason.
* Refactored provider messages and tests to use sdk.Attribute{}
* Lint fixes to use errors.Is()

references #455

Signed-off-by: Joshua Roppo <josh@akash.network>

* refactor: provider test closures

golangci lint'er does not appreciate the nested anonymous function calls
using an anonymous `test` struct. This refactor creates functions to
produce the func(t *testing.T) interface so that functionality of the
tests can still be nested for clean output, but results in more complex
code. However I do appreciate the encapsulation of the test
functionality with the types.

Signed-off-by: Joshua Roppo <josh@akash.network>

Co-authored-by: Joshua Roppo <ropes@localhost.localdomain>