feat: Makefile target matching GH linters
Provides the ability to run the same golangci-lint tests which are
triggered for every PR.

`make test-sublinters` utilizes the same config file as GH tests. Two of
the linters are disabled same as they are for GH PRs with
`continue-on-error`.

When a linter check is broken, it fails the Make target and prints cause
of failure.

@boz suggested a great Makefile trick to provide individual lint'r
execution via `make test-sublinter-<linter>`
eg: `make test-sublinter-deadcode` runs only the `deadcode` linter.

GH Test Actions now use the Makefile target to execute `make
test-sublinters` for all the linters. Unforunately fails and exits on
the first failed linter.

Signed-off-by: Josh Roppo <josh@akash.network>