feat: Close Group Message functionality
Enable closing a specific group within an existing Akash Deployment.
This will signal the provider to set the Group.State to Closed.

`akashctl tx deployment group-close --owner=? --dseq=? --gseq=?` command
added to send the `MsgCloseGroup`.

Simulation tests passing: With a prequisite in SimulateMsgCloseGroup() to skip
sending the MsgGroupClose if the Group.State is already Closed. This is
not the same behavior as the Deployment's Simulation function, but
Simulate...Group throws the handleMsgCloseGroup() ErrGroupClosed when
the simulation runs.
* Group{} ID & Validate* methods unit tested.

fixes: #642

Signed-off-by: Josh Roppo <josh@akash.network>