feat: Deployment State Indexing
Updated Deployment indexing key to enable filtering Deployments by their
State. This is hopefully to improve the performance of the deployment module's
OnEndBlock(). Provided by adding a State prefix value to the key, which is
passed to KVStorePrefixIterator(..., kvPrefix).

| Key      | Format                              |
|------|---                                      |
| Previous |               `0x01[OwnerID][DSeq]`  |
| New State Key   |  `0x01[State][OwnerID][DSeq]` |

Adding the state value behind the `deployment` prefix `0x01` value does
create the need for additional checks to be run when
GetDeployment(types.DeploymentID) is called, since it does not provide
the state, so the three potential state keys are generated and iterated
over in store.Get(key) calls, and then short circuit returns the
Deployment value.

Google's Fuzz library is imported to feed random values into tests of the
Key creating methods for durability testing.

Existing tests of the deployment module's code are fairly good, and
after finalizing the new key prefix format, they all pass.

fixes: #604

Signed-off-by: Josh Roppo <josh@akash.network>