fix: `tx` commands missing --keyring-backend
Fixes problems related to using keys not hosted in the `os` keyring
backend. This comes up with managing keys for testnets in the
`tx` subcommands. Enables management of remote resources using
local keys.

By adding the flag, `akash` can be correctly directed to reference a
`test` keyring in directories where sets of keys are created and not
stored in the OS keyring.

Signed-off-by: Josh Roppo <josh@akash.network>