docs(readme): remove stable and unstable version ambiguity
I got the impression that the latest stable version was `v0.8.0` and
latest unstable version was `v0.9.0` while looking at the README's
"Branching and Versioning" section.

But in reality the latest are `v0.30.0` and `v0.29.0` respectively at
the time of writing this commit message.
Hence adding an `e.g.` to the README page.

Signed-off-by: Khosrow Moossavi <khos2ow@gmail.com>